refactor(projects): separate project layouts and shared controls
