+++ renderer/src/views/projects/components/projectsList.component.tsx
@@ -7,638 +7,20 @@ import {
     type DragStartEvent,
 } from '@dnd-kit/react';
 import { isSortable, useSortable } from '@dnd-kit/react/sortable';
-import type {
-    CodeEditorIntegrationSettings,
-    ProjectDetails,
-    ReleaseSummary,
-} from '@shared/contracts';
-import {
-    Clock,
-    Download,
-    EllipsisVertical,
-    FlaskConical,
-    FolderOpen,
-    GripVertical,
-    ImageOff,
-    PanelTop,
-    Pin,
-    Play,
-    Settings,
-    Tag,
-    TriangleAlert,
-} from 'lucide-react';
+import type { ProjectDetails } from '@shared/contracts';
+import { GripVertical } from 'lucide-react';
 import type React from 'react';
 import { useEffect, useMemo, useRef, useState } from 'react';
-import gitIconColor from '../../../assets/icons/git_icon_color.svg';
 import githubInvertocatBlack from '../../../assets/icons/github-invertocat-black.svg';
 import githubInvertocatWhite from '../../../assets/icons/github-invertocat-white.svg';
-import { CodeEditorIntegrationIcon } from '../../../components/codeEditorIntegrationIcon.component';
-import { CopyBadge } from '../../../components/ui/copyBadge.component';
 import { Tooltip } from '../../../components/ui/tooltip.component';
 import { useTheme } from '../../../hooks/useTheme';
-import { formatRelativeTime } from '../../../i18n/relativeTime';
-import type { ProjectViewMode } from '../project-view.types';
-import {
-    getInvalidProjectTableKey,
-    type ProjectSections,
-} from '../projectsView.model';
-
-type ProjectSectionKey = 'new' | 'pinned' | 'recents';
-
-type ProjectsListProps = {
-    viewMode?: ProjectViewMode;
-    sections: ProjectSections;
-    projectGitHubUrls: ReadonlyMap<string, string>;
-    loading: boolean;
-    locale: string;
-    busyProjects: string[];
-    codeEditorSettings: CodeEditorIntegrationSettings[];
-    highlightedPinnedProjectPath: string | null;
-    pinnedReorderingDisabled: boolean;
-    onPinnedHighlightComplete: () => void;
-    onReorderPinnedProjects: (orderedProjectPaths: string[]) => Promise<void>;
-    isInstalledRelease: (version: string, mono: boolean) => boolean;
-    isProjectEditorDownloading: (project: ProjectDetails) => boolean;
-    getDownloadableProjectEditor: (
-        project: ProjectDetails,
-    ) => ReleaseSummary | undefined;
-    onInstallRequiredProjectEditor: (
-        project: ProjectDetails,
-        release: ReleaseSummary,
-    ) => void;
-    onLaunchProject: (project: ProjectDetails) => void;
-    onProjectFoldersOptions: (
-        event: React.MouseEvent,
-        project: ProjectDetails,
-    ) => void;
-    onTogglePinned: (project: ProjectDetails) => void;
-    onProjectSettings: (project: ProjectDetails) => void;
-    onProjectMoreOptions: (
-        event: React.MouseEvent,
-        project: ProjectDetails,
-    ) => void;
-    t: (key: string, options?: Record<string, unknown>) => string;
-};
-
-type ProjectListItemProps = Omit<
+import type {
+    ProjectListItemProps,
+    ProjectSectionKey,
     ProjectsListProps,
-    | 'sections'
-    | 'loading'
-    | 'highlightedPinnedProjectPath'
-    | 'pinnedReorderingDisabled'
-    | 'onPinnedHighlightComplete'
-    | 'onReorderPinnedProjects'
-> & {
-    project: ProjectDetails;
-    githubIconSrc: string;
-    sectionKey: ProjectSectionKey;
-    highlighted: boolean;
-    pinnedItemRef?: (element: HTMLLIElement | null) => void;
-    reorderHandle?: React.ReactNode;
-    reorderStateClassName?: string;
-};
-
-/**
- * Renders a project using shared status and actions in either presentation.
- * @param props - Project state, presentation and existing project handlers.
- */
-const ProjectListItem: React.FC<ProjectListItemProps> = ({
-    viewMode = 'cards',
-    project,
-    sectionKey,
-    highlighted,
-    pinnedItemRef,
-    reorderHandle,
-    reorderStateClassName = '',
-    locale,
-    busyProjects,
-    codeEditorSettings,
-    projectGitHubUrls,
-    githubIconSrc,
-    isInstalledRelease,
-    isProjectEditorDownloading,
-    getDownloadableProjectEditor,
-    onInstallRequiredProjectEditor,
-    onLaunchProject,
-    onProjectFoldersOptions,
-    onTogglePinned,
-    onProjectSettings,
-    onProjectMoreOptions,
-    t,
-}) => {
-    const editorDownloading = isProjectEditorDownloading(project);
-    const releaseInstalled = isInstalledRelease(
-        project.release.version,
-        project.release.mono,
-    );
-    const downloadableProjectEditor = getDownloadableProjectEditor(project);
-    const selectedCodeEditor = project.codeEditorId
-        ? codeEditorSettings.find(
-              (settings) => settings.integration.id === project.codeEditorId,
-          )
-        : undefined;
-    const codeEditorUnavailable = Boolean(
-        selectedCodeEditor && !selectedCodeEditor.installation,
-    );
-    const codeEditorName =
-        selectedCodeEditor?.integration.displayName ?? project.codeEditorId;
-    const codeEditorTooltip = project.codeEditorId
-        ? t(
-              codeEditorUnavailable
-                  ? 'table.codeEditorUnavailable'
-                  : 'table.codeEditorProject',
-              { editor: codeEditorName },
-          )
-        : '';
-    const editorMissing =
-        !releaseInstalled ||
-        (!project.valid && project.invalid_reason === 'missing_editor');
-    const hasWarning =
-        !project.valid || !releaseInstalled || codeEditorUnavailable;
-    const launchDisabled =
-        !project.valid || !releaseInstalled || editorDownloading;
-    const versionLabel = `${project.version}${project.release.mono ? ' (.NET)' : ''}`;
-    const isGitHubProject = projectGitHubUrls.has(project.path);
-
-    const compact = viewMode === 'list';
-    const actions = (
-        <div
-            className={`flex shrink-0 items-center gap-2 ${compact ? 'self-end' : 'self-start'}`}
-        >
-            {reorderHandle}
-            {downloadableProjectEditor && !releaseInstalled && (
-                <Tooltip placement="top" tip={t('card.installRequiredEditor')}>
-                    <button
-                        type="button"
-                        data-testid="btnInstallRequiredProjectEditor"
-                        disabled={editorDownloading}
-                        className="btn btn-ghost btn-square h-7 min-h-7 w-7 border border-warning/60 bg-base-100/20 text-warning"
-                        aria-label={t('card.installRequiredEditor')}
-                        onClick={() =>
-                            onInstallRequiredProjectEditor(
-                                project,
-                                downloadableProjectEditor,
-                            )
-                        }
-                    >
-                        {editorDownloading ? (
-                            <span className="loading loading-spinner loading-xs" />
-                        ) : (
-                            <Download size={15} />
-                        )}
-                    </button>
-                </Tooltip>
-            )}
-            <Tooltip
-                placement="top"
-                tip={t(
-                    project.pinned
-                        ? 'project.unpinProject'
-                        : 'project.pinProject',
-                    { ns: 'menus' },
-                )}
-            >
-                <button
-                    type="button"
-                    data-testid="btnToggleProjectPinned"
-                    className={`btn btn-ghost btn-square h-7 min-h-7 w-7 border bg-base-100/20 ${project.pinned ? 'border-primary/50 text-primary' : 'border-base-300'}`}
-                    aria-label={t(
-                        project.pinned
-                            ? 'project.unpinProject'
-                            : 'project.pinProject',
-                        { ns: 'menus' },
-                    )}
-                    onClick={() => onTogglePinned(project)}
-                >
-                    <Pin size={16} />
-                </button>
-            </Tooltip>
-            <Tooltip placement="top" tip={t('card.openFolders')}>
-                <button
-                    type="button"
-                    data-testid="btnProjectFolders"
-                    className="btn btn-ghost btn-square h-7 min-h-7 w-7 border border-base-300 bg-base-100/20"
-                    aria-label={t('card.openFolders')}
-                    onClick={(event) => onProjectFoldersOptions(event, project)}
-                >
-                    <FolderOpen size={16} />
-                </button>
-            </Tooltip>
-            <Tooltip placement="top" tip={t('card.projectSettings')}>
-                <button
-                    type="button"
-                    data-testid="btnProjectSettings"
-                    className="btn btn-ghost btn-square h-7 min-h-7 w-7 border border-base-300 bg-base-100/20"
-                    aria-label={t('card.projectSettings')}
-                    onClick={() => onProjectSettings(project)}
-                >
-                    <Settings size={16} />
-                </button>
-            </Tooltip>
-            <button
-                type="button"
-                data-testid="btnProjectMoreOptions"
-                onClick={(event) => onProjectMoreOptions(event, project)}
-                className="btn btn-ghost btn-square h-7 min-h-7 w-7 border border-base-300 bg-base-100/20"
-                aria-label={t('table.moreOptions', {
-                    project: project.name,
-                })}
-            >
-                <EllipsisVertical size={17} />
-            </button>
-        </div>
-    );
-    const badges = (
-        <div
-            data-testid="projectBadges"
-            className={
-                compact
-                    ? 'flex shrink-0 items-center gap-1.5'
-                    : 'flex min-w-0 flex-wrap content-start items-start gap-1.5'
-            }
-        >
-            {!compact && (
-                <Tooltip
-                    role={compact ? 'img' : undefined}
-                    ariaLabel={
-                        releaseInstalled
-                            ? t('card.godotVersion', { version: versionLabel })
-                            : t('table.invalidReasons.missingEditor')
-                    }
-                    placement="top"
-                    tip={
-                        releaseInstalled
-                            ? t('card.godotVersion', {
-                                  version: versionLabel,
-                              })
-                            : t('table.invalidReasons.missingEditor')
-                    }
-                    tone={releaseInstalled ? 'default' : 'warning'}
-                >
-                    <span
-                        className={`${compact ? 'inline-flex items-center gap-1' : 'badge badge-outline h-7 gap-1.5 px-2 text-xs'} ${releaseInstalled ? 'border-base-content/25' : 'border-warning/60 text-warning'}`}
-                    >
-                        {editorDownloading ? (
-                            <span className="loading loading-spinner loading-xs" />
-                        ) : releaseInstalled ? (
-                            !compact && <Tag size={13} />
-                        ) : (
-                            <TriangleAlert size={compact ? 16 : 13} />
-                        )}
-                        {!compact && <span>{versionLabel}</span>}
-                        {project.release.prerelease && (
-                            <FlaskConical
-                                size={compact ? 16 : 12}
-                                className="text-secondary"
-                            />
-                        )}
-                    </span>
-                </Tooltip>
-            )}
-
-            {project.codeEditorId && (
-                <Tooltip
-                    placement="top"
-                    role={compact ? 'img' : undefined}
-                    ariaLabel={codeEditorTooltip}
-                    tip={codeEditorTooltip}
-                    tone={codeEditorUnavailable ? 'warning' : 'default'}
-                >
-                    <span
-                        className={`${compact ? 'inline-flex items-center gap-1' : 'badge badge-outline h-7 gap-1.5 px-2 text-xs'} ${codeEditorUnavailable ? 'border-warning/60 text-warning' : 'border-base-content/25'}`}
-                    >
-                        {codeEditorUnavailable ? (
-                            <TriangleAlert
-                                size={compact ? 16 : 13}
-                                className="stroke-warning"
-                            />
-                        ) : (
-                            <CodeEditorIntegrationIcon
-                                integrationId={project.codeEditorId}
-                                className={compact ? 'size-[16px]' : 'size-3.5'}
-                            />
-                        )}
-                        <span
-                            className={
-                                compact ? 'sr-only' : 'max-w-48 truncate'
-                            }
-                        >
-                            {codeEditorName}
-                        </span>
-                    </span>
-                </Tooltip>
-            )}
-
-            {project.withGit && (
-                <Tooltip
-                    placement="top"
-                    tip={t(
-                        isGitHubProject
-                            ? 'table.githubProject'
-                            : 'table.gitProject',
-                    )}
-                    tone="default"
-                >
-                    <span
-                        className={
-                            compact
-                                ? 'inline-flex items-center'
-                                : 'badge badge-outline h-7 gap-1.5 border-base-content/25 px-2 text-xs'
-                        }
-                    >
-                        {isGitHubProject ? (
-                            <img
-                                src={githubIconSrc}
-                                className={
-                                    compact ? 'size-[16px]' : 'h-3.5 w-3.5'
-                                }
-                                alt=""
-                                aria-hidden="true"
-                                data-testid="githubProjectIcon"
-                            />
-                        ) : (
-                            <img
-                                src={gitIconColor}
-                                className={
-                                    compact ? 'size-[16px]' : 'h-3.5 w-3.5'
-                                }
-                                alt=""
-                                data-testid="gitProjectIcon"
-                            />
-                        )}
-                        <span className={compact ? 'sr-only' : undefined}>
-                            {isGitHubProject ? 'GitHub' : 'Git'}
-                        </span>
-                    </span>
-                </Tooltip>
-            )}
-
-            {project.open_windowed && (
-                <Tooltip
-                    placement="top"
-                    tip={t('table.windowedMode')}
-                    tone="default"
-                >
-                    <span
-                        className={
-                            compact
-                                ? 'inline-flex items-center'
-                                : 'badge badge-outline h-7 gap-1.5 border-base-content/25 px-2 text-xs'
-                        }
-                    >
-                        <PanelTop size={compact ? 16 : 13} />
-                        <span className={compact ? 'sr-only' : undefined}>
-                            {t('card.windowed')}
-                        </span>
-                    </span>
-                </Tooltip>
-            )}
-        </div>
-    );
-    const lastOpened = project.last_opened
-        ? t('card.opened', {
-              age: formatRelativeTime(project.last_opened, locale),
-          })
-        : t('card.notOpened');
-
-    const compactLaunchButton = (
-        <button
-            type="button"
-            data-testid="btnLaunchCompactProject"
-            disabled={launchDisabled}
-            aria-label={t('view.openProject', {
-                project: project.name,
-            })}
-            onClick={() => onLaunchProject(project)}
-            className="group flex h-[24px] min-w-0 items-center gap-2 rounded-sm text-left focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary disabled:opacity-45 disabled:cursor-not-allowed"
-        >
-            <span className="absolute left-0 top-1/2 flex size-[48px] -translate-y-1/2 items-center justify-center overflow-hidden rounded bg-base-content/8 group-focus-visible:outline-2 group-focus-visible:outline-primary">
-                {project.icon_path ? (
-                    <img
-                        src={project.icon_path}
-                        className="size-full object-contain"
-                        alt=""
-                    />
-                ) : (
-                    <ImageOff className="size-5 stroke-base-content/30" />
-                )}
-            </span>
-            <span className="truncate text-[16px] font-semibold underline-offset-2 group-hover:underline">
-                {project.name}
-            </span>
-        </button>
-    );
-
-    return (
-        <li
-            ref={pinnedItemRef}
-            tabIndex={sectionKey === 'pinned' ? -1 : undefined}
-            className={`relative overflow-hidden ${compact ? 'flex min-h-[68px] items-center border-b border-base-300 px-3 py-1.5 pl-5 hover:bg-base-content/[0.03]' : 'rounded-lg border border-base-300 bg-base-200/35 p-4 pl-5 shadow-sm hover:bg-base-200/55'} transition-colors motion-reduce:transition-none hover:border-base-content/20 ${highlighted ? 'project-pin-highlight' : ''} ${reorderStateClassName}`}
-            data-project-view={viewMode}
-            data-project-path={project.path}
-            data-project-section={sectionKey}
-        >
-            <div
-                className={`absolute inset-y-3 left-3 w-1 rounded-full ${hasWarning ? 'bg-warning' : 'bg-base-content/15'}`}
-                aria-hidden="true"
-            />
-            {busyProjects.includes(project.path) && (
-                <div className="absolute inset-0 z-10 flex items-center justify-center rounded-lg bg-black/55">
-                    <div className="loading loading-bars" />
-                </div>
-            )}
-
-            {compact ? (
-                <div className="flex w-full min-w-0 items-center gap-2 pl-2">
-                    <div className="grid min-w-0 flex-1 grid-cols-[minmax(0,1fr)_minmax(12rem,auto)] items-stretch gap-4">
-                        <div className="relative flex min-h-[48px] min-w-0 flex-col justify-center pl-[60px]">
-                            <div className="flex min-w-0 items-center gap-2">
-                                {launchDisabled ? (
-                                    compactLaunchButton
-                                ) : (
-                                    <Tooltip
-                                        tip={t('card.editInGodot')}
-                                        placement="top"
-                                        className="min-w-0"
-                                    >
-                                        {compactLaunchButton}
-                                    </Tooltip>
-                                )}
-                                {!project.valid &&
-                                    project.invalid_reason !==
-                                        'missing_editor' && (
-                                        <Tooltip
-                                            placement="top"
-                                            tip={t(
-                                                getInvalidProjectTableKey(
-                                                    project,
-                                                ),
-                                            )}
-                                            role="img"
-                                            ariaLabel={t(
-                                                getInvalidProjectTableKey(
-                                                    project,
-                                                ),
-                                            )}
-                                            tone="warning"
-                                        >
-                                            <TriangleAlert className="size-[16px] shrink-0 text-warning" />
-                                        </Tooltip>
-                                    )}
-                                {badges}
-                            </div>
-                            <CopyBadge
-                                value={project.path}
-                                label={t('common:buttons.copyPath')}
-                                copiedLabel={t('common:success')}
-                                className="self-start hover:bg-base-100"
-                                data-testid={`btnCopyProjectPath_${sectionKey}_${project.path}`}
-                            />
-                        </div>
-                        <div className="flex min-h-[48px] min-w-0 max-w-80 flex-col items-end self-stretch justify-between gap-1">
-                            {actions}
-                            <div className="flex w-full min-w-0 items-center justify-end gap-2 text-xs text-base-content/55">
-                                <Tooltip
-                                    tip={
-                                        editorMissing
-                                            ? t(
-                                                  'table.invalidReasons.missingEditor',
-                                              )
-                                            : t('card.godotVersion', {
-                                                  version: versionLabel,
-                                              })
-                                    }
-                                    tone={editorMissing ? 'warning' : 'default'}
-                                    placement="top"
-                                    className="min-w-0"
-                                >
-                                    <span
-                                        data-testid="compactProjectEditorVersion"
-                                        className={`flex min-w-0 items-center gap-1 ${editorMissing ? 'text-warning' : ''}`}
-                                    >
-                                        {editorMissing && (
-                                            <>
-                                                <TriangleAlert
-                                                    size={16}
-                                                    className="shrink-0"
-                                                    aria-hidden="true"
-                                                />
-                                                <span className="sr-only">
-                                                    {t(
-                                                        'table.invalidReasons.missingEditor',
-                                                    )}
-                                                </span>
-                                            </>
-                                        )}
-                                        {editorDownloading && (
-                                            <span className="loading loading-spinner loading-xs shrink-0" />
-                                        )}
-                                        {project.release.prerelease && (
-                                            <FlaskConical
-                                                size={16}
-                                                className="shrink-0 text-secondary"
-                                            />
-                                        )}
-                                        <span className="min-w-0 truncate">
-                                            {versionLabel}
-                                        </span>
-                                    </span>
-                                </Tooltip>
-                                <span
-                                    aria-hidden="true"
-                                    className="h-3 border-l border-base-300"
-                                />
-                                <Clock
-                                    size={12}
-                                    className="shrink-0"
-                                    aria-hidden="true"
-                                />
-                                <span className="min-w-0 truncate">
-                                    {lastOpened}
-                                </span>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            ) : (
-                <div className="flex min-w-0 flex-col gap-4 pl-2">
-                    <div className="grid min-w-0 grid-cols-[48px_minmax(0,1fr)_auto] items-center gap-4">
-                        <div className="flex size-12 shrink-0 items-center justify-center overflow-hidden rounded-lg bg-base-content/8">
-                            {project.icon_path ? (
-                                <img
-                                    src={project.icon_path}
-                                    className="h-full w-full object-contain"
-                                    alt=""
-                                />
-                            ) : (
-                                <ImageOff className="h-6 w-6 stroke-base-content/30" />
-                            )}
-                        </div>
-
-                        <div className="flex min-w-0 flex-col gap-1">
-                            <div className="flex min-w-0 items-center gap-2">
-                                {!project.valid && (
-                                    <Tooltip
-                                        placement="top"
-                                        tip={t(
-                                            getInvalidProjectTableKey(project),
-                                        )}
-                                        tone="warning"
-                                    >
-                                        <TriangleAlert className="size-5 shrink-0 stroke-warning" />
-                                    </Tooltip>
-                                )}
-                                <h3 className="truncate text-xl font-semibold leading-tight text-base-content">
-                                    {project.name}
-                                </h3>
-                            </div>
-                            <CopyBadge
-                                value={project.path}
-                                label={t('common:buttons.copyPath')}
-                                copiedLabel={t('common:success')}
-                                className="max-w-full self-start rounded-md bg-transparent px-0 text-base-content/55"
-                                data-testid={`btnCopyProjectPath_${sectionKey}_${project.path}`}
-                            />
-                        </div>
-
-                        {actions}
-                    </div>
-
-                    <div className="grid min-w-0 grid-cols-[minmax(0,1fr)_auto] items-start gap-6">
-                        {badges}
-
-                        <div
-                            data-testid="projectLaunchActions"
-                            className="flex min-w-36 shrink-0 items-center gap-2"
-                        >
-                            <p className="whitespace-nowrap text-sm text-base-content/55">
-                                {project.last_opened
-                                    ? t('card.opened', {
-                                          age: formatRelativeTime(
-                                              project.last_opened,
-                                              locale,
-                                          ),
-                                      })
-                                    : t('card.notOpened')}
-                            </p>
-
-                            <button
-                                type="button"
-                                data-testid="btnEditProjectInGodot"
-                                disabled={launchDisabled}
-                                onClick={() => onLaunchProject(project)}
-                                className="btn btn-primary btn-sm min-w-32 gap-2 rounded-md"
-                            >
-                                <Play size={16} />
-                                {t('card.editInGodot')}
-                            </button>
-                        </div>
-                    </div>
-                </div>
-            )}
-        </li>
-    );
-};
+} from './project-list.types';
+import { ProjectListItem } from './project-list-item.component';
 
 type SortablePinnedProjectItemProps = Omit<
     ProjectListItemProps,
